requirements.txt
flake8
black
django
flake8-blind-except 
flake8-builtins 
flake8-logging-format 
flake8-commas 
flake8-print 
pep8-naming 
flake8-cognitive-complexity 
flake8-variables-names 
flake8-comprehensions 
flake8-django
flake8-eradicate
django-debug-toolbar
phonenumbers
django-phonenumber-field
pillow
django-crispy-forms